backend/requirements.txt
# Django y Django REST Framework
Django==5.2.8
djangorestframework==3.16.1

# Autenticación JWT
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# CORS Headers
django-cors-headers==4.9.0

# Base de datos
psycopg2-binary==2.9.11
dj-database-url==3.0.1

# Variables de entorno
python-decouple==3.8

# Servidor WSGI para producción
gunicorn==21.2.0

# Testing
coverage==7.12.0

# Dependencias de Django
asgiref==3.11.0
sqlparse==0.5.3
tzdata==2025.2

